Always show menu button
If I've done several searches, getting back to the home screen will take
too many clicks using the back button. When
`enable-menu-with-back-views` is set to false, there is no menu button,
and it's also not possible to swipe from left to open the menu, so I set
it to true instead.